it3 jason web token codice setup

+++ Codice/HackHub/src/main/java/com/unicam/hackhub/Controller/StaffController.java
@@ -62,7 +62,7 @@ public class StaffController {
 
     @PostMapping("/valuta")
     public ResponseEntity<Object> valuta(
-            @RequestParam ("gdc") Integer giudiceId,
+            @RequestParam ("gdc") Long giudiceId,
             @RequestBody ValutazioneInfo infoVal,
             @RequestParam ("tm") String nomeTeam
     ){
